Finishing the blog api

+++ backend/blog/views.py
@@ -1,13 +1,17 @@
+from rest_framework.pagination import PageNumberPagination
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
 from rest_framework import permissions
-from rest_framework.serializers import Serializer
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.views import APIView
 from rest_framework.generics import ListAPIView, RetrieveAPIView
+from rest_framework.filters import SearchFilter, OrderingFilter
+
 from accounts.models import UserAccount
 from .models import BlogPost
 from .serializers import BlogPostSerializer
+
 
 
 class BlogPostListView(ListAPIView):
@@ -15,13 +19,17 @@ class BlogPostListView(ListAPIView):
     serializer_class = BlogPostSerializer
     lookup_field = 'slug'
     permission_classes = (permissions.AllowAny, )
+    pagination_class = PageNumberPagination
+
+    filter_backends = (SearchFilter, OrderingFilter)
+    search_fields = ('title', 'excerpt', 'content', 'author__username')
 
 
 class BlogPostDetailView(RetrieveAPIView):
     queryset = BlogPost.objects.order_by('-date_created')
     serializer_class = BlogPostSerializer
     lookup_field = 'slug'
-    permission_classes = (permissions.AllowAny, )
+    permission_classes = (permissions.IsAuthenticated, )
 
 
 class BlogPostFeaturedView(ListAPIView):
@@ -47,12 +55,17 @@ class BlogPostCategoryView(APIView):
 
 
 @api_view(['PUT', ])
-def api_update_view(request, slug):
+@permission_classes((IsAuthenticated,))
+def api_update_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return  Response({'response': "You do not have permission to edit this post."})
 
     if request.method == "PUT":
         serializer = BlogPostSerializer(blog_post, data=request.data)
@@ -65,11 +78,16 @@ def api_update_view(request, slug):
 
 
 @api_view(['DELETE', ])
-def api_update_blog_view(request, slug):
+@permission_classes((IsAuthenticated,))
+def api_delete_blog_view(request, slug):
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You do not have permission to delete this post."})
 
     if request.method == "DELETE":
         operation = blog_post.delete()
@@ -82,9 +100,10 @@ def api_update_blog_view(request, slug):
 
 
 @api_view(['POST', ])
+@permission_classes((IsAuthenticated,))
 def api_create_blog_view(request):
 
-    account = UserAccount.objects.get(pk=1)
+    account = request.user
 
     blog_post = BlogPost(author=account)
